Improves compile timing and allow modification of parameters online
