+++ robots/xuanwu/cfg/Planner.cfg
@@ -6,31 +6,31 @@ from dynamic_reconfigure.parameter_generator_catkin import *
 gen = ParameterGenerator()
 
 #       Name                 Type      Level  Description                       Default  Min   Max
-gen.add("control_rate",      double_t, 0,     "MPC Control Rate (Hz)",          20.0,    10.0, 100.0)
-gen.add("horizon",           int_t,    0,     "Prediction Horizon (steps)",     20,      5,    50)
-gen.add("lookahead_steps",   int_t,    0,     "Lookahead Step (for Controller)", 3,      1,    10)
+gen.add("control_rate",      double_t, 0,     "MPC Control Rate (Hz)",          20.0,    10.0, 400.0)
+gen.add("horizon",           int_t,    0,     "Prediction Horizon (steps)",     20,      5,    100)
+gen.add("lookahead_steps",   int_t,    0,     "Lookahead Step (for Controller)", 0,      0,    100)
 
 # --- Q Weights (State Penalty) ---
 group_pos = gen.add_group("Weights Position")
-group_pos.add("Q_pos_x", double_t, 0, "Weight Pos X", 10.0, 0.0, 1000.0)
-group_pos.add("Q_pos_y", double_t, 0, "Weight Pos Y", 10.0, 0.0, 1000.0)
-group_pos.add("Q_pos_z", double_t, 0, "Weight Pos Z", 10.0, 0.0, 1000.0)
+group_pos.add("Q_pos_x", double_t, 0, "Weight Pos X", 10.0, 0.0, 10000.0)
+group_pos.add("Q_pos_y", double_t, 0, "Weight Pos Y", 10.0, 0.0, 10000.0)
+group_pos.add("Q_pos_z", double_t, 0, "Weight Pos Z", 10.0, 0.0, 10000.0)
 
 group_att = gen.add_group("Weights Attitude")
-group_att.add("Q_att_r", double_t, 0, "Weight Roll",  1.0, 0.0, 1000.0)
-group_att.add("Q_att_p", double_t, 0, "Weight Pitch", 1.0, 0.0, 1000.0)
-group_att.add("Q_att_y", double_t, 0, "Weight Yaw",   1.0, 0.0, 1000.0)
+group_att.add("Q_att_r", double_t, 0, "Weight Roll",  1.0, 0.0, 10000.0)
+group_att.add("Q_att_p", double_t, 0, "Weight Pitch", 1.0, 0.0, 10000.0)
+group_att.add("Q_att_y", double_t, 0, "Weight Yaw",   1.0, 0.0, 10000.0)
 
 group_vel = gen.add_group("Weights Velocity")
-group_vel.add("Q_vel_x", double_t, 0, "Weight Vel X", 1.0, 0.0, 1000.0)
-group_vel.add("Q_vel_y", double_t, 0, "Weight Vel Y", 1.0, 0.0, 1000.0)
-group_vel.add("Q_vel_z", double_t, 0, "Weight Vel Z", 1.0, 0.0, 1000.0)
+group_vel.add("Q_vel_x", double_t, 0, "Weight Vel X", 1.0, 0.0, 10000.0)
+group_vel.add("Q_vel_y", double_t, 0, "Weight Vel Y", 1.0, 0.0, 10000.0)
+group_vel.add("Q_vel_z", double_t, 0, "Weight Vel Z", 1.0, 0.0, 10000.0)
 
 group_omega = gen.add_group("Weights Angular Velocity")
-group_omega.add("Q_omega", double_t, 0, "Weight Omega (All axes)", 1.0, 0.0, 1000.0)
+group_omega.add("Q_omega", double_t, 0, "Weight Omega (All axes)", 1.0, 0.0, 10000.0)
 
 # --- R Weights (Input Penalty) ---
 group_input = gen.add_group("Weights Input")
-group_input.add("R_thrust", double_t, 0, "Weight Thrust (per motor)", 1.0, 0.0, 1000.0)
+group_input.add("R_thrust", double_t, 0, "Weight Thrust (per motor)", 1.0, 0.0, 10000.0)
 
 exit(gen.generate(PACKAGE, "xuanwu", "Planner"))
